Changed language to english
After some soulsearching I decided to change the language (back) to English. The documentation and comments will still be in German, but method names and variables are completely in English now.

Besides that, a lot of imprvoements have been made in many areas and the framework is coming together niecely.